job_openings/requirements.txt

requests
python-dotenv
numpy

google-genai

streamlit
crewai
